Pull out target-land annotations into a seperate project
