+++ targetutils/build.sbt
@@ -2,6 +2,6 @@ organization := "edu.berkeley.cs"
 
 version := "1.0-SNAPSHOT"
 
-name := "midas"
+name := "midas-targetutils"
 
 scalaVersion := "2.12.4"
